requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
fabric==2.7.0
idna==3.3
invoke==1.7.0
paramiko==2.10.3
pathlib2==2.3.7.post1
pika==1.2.0
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pync==2.0.3
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
sshtunnel==0.4.0
toml==0.10.2
urllib3==1.26.9

watchdog~=2.1.8